Refactor: ParticipantsList 컴포넌트 분리
